moved TimerFunctions to utils folder
