+++ emotion_tracker/src/utils/TimerFunctions.js
